[SID-1438] Jump page: password recovery (#210)
* Fix the config to use a stable path to react primitives

* Render the recovery state based on challenge types

* Implement the full recovery flow

* Leave a TODO for the incorrect success message after verifying the handle

* Fix the jump page CLI

* Build the jump page

* Use the absolute path to the dist folder

* Make the jump-page package public to simplify ttesting

* Resolve todos

* Refactor the form authenticating state to prevent duplication

* Refactor - remove listener for reset state

* Use the OTP Input component from the primitives package

* Split the input components into multiple source files

* Publish the react-primitives containing just the source files

* Publish the react-primitives with the source only

* Set the correct package version

* Remove the jump page as it was moved to ng-evangelion
