--- apps/jump-page/.vscode/extensions.json
@@ -1,4 +1,0 @@
-{
-  "recommendations": ["astro-build.astro-vscode"],
-  "unwantedRecommendations": []
-}
