--- apps/jump-page/CHANGELOG.md
@@ -1,37 +1,0 @@
-# @slashid/jump-page
-
-## 0.0.6
-
-### Patch Changes
-
-- Updated dependencies [64212fd]
-  - @slashid/react-primitives@0.1.4
-
-## 0.0.5
-
-### Patch Changes
-
-- Updated dependencies [21c2948]
-  - @slashid/react-primitives@0.1.3
-
-## 0.0.4
-
-### Patch Changes
-
-- Updated dependencies [0e2fb6b]
-  - @slashid/react-primitives@0.1.2
-
-## 0.0.3
-
-### Patch Changes
-
-- Updated dependencies [0e4746c]
-  - @slashid/react-primitives@0.1.1
-
-## 0.0.2
-
-### Patch Changes
-
-- c1cf068: Update dependencies
-- Updated dependencies [c1cf068]
-  - @slashid/react-primitives@0.1.0
