--- apps/jump-page/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@slashid/jump-page",
-  "type": "module",
-  "version": "0.0.6",
-  "private": false,
-  "publishConfig": {
-    "access": "restricted"
-  },
-  "bin": {
-    "sid-jump-cli": "./cli.js"
-  },
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "test": "astro check",
-    "build": "astro build",
-    "preview": "astro preview",
-    "astro": "astro"
-  },
-  "dependencies": {
-    "@astrojs/check": "^0.3.1",
-    "@astrojs/react": "^3.0.5",
-    "@sentry/react": "^7.81.1",
-    "@slashid/react-primitives": "workspace:*",
-    "@slashid/slashid": "3.17.1",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
-    "astro": "^3.5.5",
-    "commander": "^11.1.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "devDependencies": {
-    "@vanilla-extract/css": "^1.9.2",
-    "@vanilla-extract/vite-plugin": "^3.9.2"
-  }
-}
