--- apps/jump-page/src/components/README.md
@@ -1,3 +1,0 @@
-# Components
-
-This folder should only contain components meant to render server side and then hydrate on the client side.
