--- apps/jump-page/src/components/app/index.ts
@@ -1,1 +1,0 @@
-export * from "./app.component";
