--- apps/jump-page/src/components/flow/flow.loader.tsx
@@ -1,9 +1,0 @@
-import { Circle, Spinner } from "@slashid/react-primitives";
-
-export function Loader() {
-  return (
-    <Circle>
-      <Spinner />
-    </Circle>
-  );
-}
