--- apps/jump-page/src/components/flow/index.ts
@@ -1,3 +1,0 @@
-export * from "./flow.component";
-export * from "./flow.types";
-export * from "./flow.error";
