--- apps/jump-page/src/layouts/README.md
@@ -1,3 +1,0 @@
-# Layouts
-
-This folder should only contain [Astro layouts](https://docs.astro.build/en/core-concepts/layouts/).
