--- apps/jump-page/src/layouts/page.astro
@@ -1,33 +1,0 @@
----
-interface Props {
-	title: string;
-}
-
-const { title } = Astro.props;
----
-
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8" />
-        <meta content="width=device-width, initial-scale=1" name="viewport" />
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-        <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500&display=swap" rel="stylesheet" />
-		<link rel="icon" href="/favicon.png">
-		<title>{title}</title>        
-    </head>
-
-    <body>
-		<slot />
-    </body>
-</html>
-
-<style>
-	body {
-		width: 100%;
-		height: 100vh;
-		margin: 0;
-		padding: 0;
-	}	
-</style>
