--- apps/jump-page/src/pages/index.astro
@@ -1,8 +1,0 @@
----
-import Page from "../layouts/page.astro";
-import { App } from "../components/app";
----
-
-<Page title="/id | Jump page">
-    <App client:load/> 
-</Page>
