+++ packages/react-primitives/src/components/input/input.password.tsx
@@ -1,0 +1,9 @@
+import { BaseInput, BaseInputProps } from "./input.base";
+
+export type PasswordInputProps = Omit<BaseInputProps, "type">;
+
+export const PasswordInput: React.FC<PasswordInputProps> = (props) => {
+  // TODO add the eye icon to toggle between show password and hide password states
+  // TODO add the error prop to show the error state (red border)
+  return <BaseInput {...props} type={"password"} />;
+};
