+++ packages/react-primitives/src/main.ts
@@ -5,7 +5,13 @@ import {
 } from "./components/theme-root";
 import { Button } from "./components/button";
 import { LinkButton } from "./components/button/link-button";
-import { Input, PhoneInput, Flag } from "./components/input";
+import {
+  Input,
+  PhoneInput,
+  PasswordInput,
+  Flag,
+  OtpInput,
+} from "./components/input";
 import { Divider } from "./components/divider";
 import { Accordion } from "./components/accordion";
 import { Dialog, DialogProps } from "./components/dialog";
@@ -36,6 +42,8 @@ export {
   LinkButton,
   Input,
   PhoneInput,
+  OtpInput,
+  PasswordInput,
   Divider,
   Accordion,
   Dialog,
