+++ packages/react/src/components/form/authenticating/index.tsx
@@ -1,33 +1,17 @@
-import { Factor } from "@slashid/slashid";
-
-import {
-  isFactorEmailLink,
-  isFactorOTP,
-  isFactorPassword,
-  isFactorSmsLink,
-} from "../../../domain/handles";
+import { isFactorOTP, isFactorPassword } from "../../../domain/handles";
 import { Text } from "../../text";
 
-import { EmailIcon, SmsIcon, Loader } from "./icons";
 import { getAuthenticatingMessage } from "./messages";
 import { OTPState } from "./otp";
 import { PasswordState } from "./password";
 import { Props } from "./authenticating.types";
-import { BackButton, RetryPrompt } from "./authenticating.components";
+import {
+  BackButton,
+  FactorIcon,
+  RetryPrompt,
+} from "./authenticating.components";
 
 import * as styles from "./authenticating.css";
-
-const FactorIcon = ({ factor }: { factor: Factor }) => {
-  if (isFactorEmailLink(factor)) {
-    return <EmailIcon />;
-  }
-
-  if (isFactorSmsLink(factor)) {
-    return <SmsIcon />;
-  }
-
-  return <Loader />;
-};
 
 const LoadingState = ({ flowState }: Props) => {
   const factor = flowState.context.config.factor;
