+++ packages/react/src/components/form/authenticating/validation.ts
@@ -33,21 +33,6 @@ export const isValidOTPCode: ValidatorFn<string> = (value) => {
   return true;
 };
 
-// TODO copied from the core SDK - consider exposing from there
-export const PasswordValidationRuleName = {
-  Length: "length",
-  PasswordVariants: "password_variants",
-  AdminVariants: "admin_variants",
-  UserVariants: "user_variants",
-  AlphanumericSequences1: "alphanumeric_sequences_1",
-  AlphanumericSequences2: "alphanumeric_sequences_2",
-  NumericSequencesAscending: "numeric_sequences_ascending",
-  NumericSequencesDescending: "numeric_sequences_descending",
-  NumericSubsequencesAscending: "numeric_subsequences_ascending",
-  NumericSubsequencesDescending: "numeric_subsequences_descending",
-  CommonPasswordXkcd: "common_password_xkcd",
-} as const;
-
 export function getValidationMessageKey(
   errorEvent: InvalidPasswordSubmittedEvent
 ): TextConfigKey {
